+++ src/GUI/GUI_TabPane.h
@@ -48,8 +48,8 @@ public:
 
 	virtual	void			ReceiveMessage(
 									GUI_Broadcaster *		inSrc,
-									long					inMsg,
-									long					inParam);
+									intptr_t				inMsg,
+									intptr_t				inParam);
 			
 private:
