+++ src/WEDWindows/WED_DocumentWindow.h
@@ -54,8 +54,8 @@ public:
 
 	virtual	void	ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							long					inMsg,
-							long					inParam);
+							intptr_t				inMsg,
+							intptr_t				inParam);
 
 	virtual	bool	Closed(void);
